+++ src/models/Commands/GameMenuCommands.java
@@ -54,8 +54,7 @@ public enum GameMenuCommands {
     COLLECT_PRODUCE("^\\s*collect\\s+produce\\s+-n\\s+(?<name>.*\\S)\\s*$"),
     SELL_ANIMAL("^\\s*sell\\s+animal\\s+-n\\s+(?<name>.*\\S)\\s*$"),
     FISHING("^\\s*fishing\\s+-p\\s+(?<fishingPole>.*\\S)\\s*$"),
-    ARTISAN_USE("^\\s*artisan\\s+use\\s+(?<artisanName>.*\\S)\\s+(?<item1Name>.*\\S)\\s*$"),
-    ARTISAN_USE_FIXED("^\\s*artisan\\s+use\\s+Fish_Smoker\\s+Salmon\\s+Coal\\s*$"),
+    ARTISAN_USE("^\\s*artisan\\s+use\\s+(?<artisanName>.*\\S)\\s+(?<itemName>.*\\S)\\s*$"),
     ARTISAN_GET("^\\s*artisan\\s+get\\s+(?<artisanName>.*\\S)\\s*$"),
     SHOW_ALL_PRODUCTS("^\\s*show\\s+all\\s+products\\s*$"),
     SHOW_ALL_AVAILABLE_PRODUCTS("^\\s*show\\s+all\\s+available\\s+products\\s*$"),
