+++ src/models/Items/IndustrialProducts/IndustrialProduct.java
@@ -2,15 +2,15 @@ package models.Items.IndustrialProducts;
 
 import models.Items.Item;
 
-public class IndustrialProduct extends Item{
-    private final IndustrialProductType type;
+public class IndustrialProduct extends Item {
+    private final CraftingRecipe type;
 
-    public IndustrialProduct(int count, IndustrialProductType type) {
+    public IndustrialProduct(int count, CraftingRecipe type) {
         super(count, type.getName());
         this.type = type;
     }
 
-    public IndustrialProductType getType() {
+    public CraftingRecipe getType() {
         return type;
     }
