+++ src/models/Players/Player.java
@@ -9,9 +9,9 @@ import models.Invetory.ShippingBin;
 import models.Invetory.ShippingBinType;
 import models.Invetory.TrashCan;
 import models.Invetory.TrashCanType;
-import models.Items.Products.CraftingRecipe;
 import models.Items.Item;
 import models.Items.Foods.FoodType;
+import models.Items.IndustrialProducts.CraftingRecipe;
 import models.Items.Products.ForgingSeed;
 import models.Items.Products.ForgingSeedType;
 import models.Items.Tools.*;
